move some functions to hqcase.dbaccessors
