+++ corehq/apps/domain/calculations.py
@@ -4,6 +4,7 @@ from dateutil.relativedelta import relativedelta
 
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext as _
+from corehq.apps.hqcase.dbaccessors import get_number_of_cases_in_domain
 from corehq.util.dates import iso_string_to_datetime
 from corehq.apps.app_manager.models import ApplicationBase
 from corehq.apps.users.util import WEIRD_USER_IDS
@@ -72,8 +73,7 @@ def active_mobile_users(domain, *args):
 
 
 def cases(domain, *args):
-    row = get_db().view("hqcase/types_by_domain", startkey=[domain], endkey=[domain, {}]).one()
-    return row["value"] if row else 0
+    return get_number_of_cases_in_domain(domain)
 
 
 def cases_in_last(domain, days):
