+++ corehq/apps/hqcase/utils.py
@@ -4,6 +4,7 @@ from xml.etree import ElementTree
 from couchdbkit import ResourceNotFound
 from django.core.files.uploadedfile import UploadedFile
 from casexml.apps.phone.xml import get_case_xml
+from corehq.apps.hqcase.dbaccessors import get_case_ids_in_domain
 from dimagi.utils.couch.database import iter_docs
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.models import CommCareCase
@@ -135,16 +136,6 @@ def get_case_by_identifier(domain, identifier):
         pass
 
     return None
-
-
-def get_case_ids_in_domain(domain, type=None):
-    type_key = [type] if type else []
-    return [res['id'] for res in CommCareCase.get_db().view('hqcase/types_by_domain',
-        startkey=[domain] + type_key,
-        endkey=[domain] + type_key + [{}],
-        reduce=False,
-        include_docs=False,
-    )]
 
 
 def get_cases_in_domain(domain, type=None):
